src: Use CryptographicOperations.ZeroMemory().
I recently discovered there was a dedicated function for zeroing memory in .NET (https://github.com/dotnet/runtime/discussions/48697).

In an ideal world, I'd pin arrays to prevent copies being made in memory. However, I'm not going to do this because it would make the code far less readable, and it's not realistically possible to implement this technique everywhere.