src: Add non-interactive password entry.
The user can randomly generate a passphrase by specifying -p:" ".

Planned support for reading a password from a text file was removed due to ambiguity and to discourage people storing passwords in unencrypted files.

Other changes include a space before the 'Incorrect password' error when decrypting a private key, the authenticated comment for signatures will not be displayed if it's empty, and the encryption/decryption validation has been improved slightly.