FileHandling.cs: Don't increment file number.
If a file name of 'name (1)' was stored encrypted, it would get restored to 'name (2)' before instead of 'name (1) (2)'.

This number approach is quite a hassle, but I'm not sure what else could be done. With encrypted file names, you can't just error if the file already exists.

Also, it's still not right for 'name.jpg (2)', which gets restored to 'name (2).jpg (2)' if 'name.jpg (2)' exists.