src: Throw exceptions during validation.
You can argue if (!validUserInput) { return; } is readable, but I think it looks horrible and is something I hate about Go.

This is a decent compromise because it's a unique exception and is clearer than returning nothing.

The downside at the moment is that an exception for reading the private key file hides a potential error message for a public key and/or pre-shared key. So, it may need to be changed.