src: Try to optimise FileStream writing.
It's pretty shocking how there's no official guidance on FileStream parameters. It's basically a guessing game.

Preallocating seems to do nothing in my benchmarks but is a recommended step (new to .NET 6). I've also read .NET team members saying to use a buffer size of 0, but that's slower in my benchmarks like with reading. Doing SetLength as well as preallocating may be worth investigating more. Then the user buffer size should be larger ideally to reduce system calls, but I want to encrypt/decrypt in 16 KiB chunks, so I can't change that easily. That makes a big difference though.