src: Screw it; time to use ChaCha20-Poly1305.
I've been extremely indecisive on a couple of design decisions because I want this to be a stable format. This one is particularly tricky. People want me to use ChaCha20-Poly1305, and I want to use it too due to the performance and standardisation. However, like most AEADs, it's not committing. In other words, these AEADs are poorly designed in a way reminiscent of length extension attacks. Without adding commitment, I'm knowingly weakening the protocol, opening doors to potential attacks.

Due to the padding fix, this should currently be CMT-1 secure. It commits to the key and nonce. However, it hasn't been investigated as much as the UtC transform and adds 32 bytes to each ciphertext. Furthermore, it means you can't just use any old ChaCha20-Poly1305 implementation; you need to combine ChaCha20 and Poly1305 yourself.

The other problem is that ChaCha20-BLAKE2b is CMT-4 secure. How bad is only having CMT-1 security? It's not very clear, but it's certainly not ideal because you'd expect an AEAD to commit to all the inputs. Making it CMT-4 secure is possible by including the associated data in the key derivation, but it gets real ugly (e.g. you can't just use Argon2 anymore).

Then do I add commitment to each chunk or only the header? Again, it's not clear. Only to the header would reduce storage overhead and be enough to prevent partitioning oracle attacks but certainly seems unwise when using a separate file key for the payload.