src: Use ephemeral public key in password/PSK key derivation.
Instead of using the ephemeral public key as associated data to detect tampering, it's now used as the BLAKE2b message for key derivation. Since the ephemeral public key is actually random bytes rather than a public key, this provides an extra 256-bit salt.

Furthermore, instead of peppering a password before passing it to Argon2, the pepper is now used as part of the input keying material for key derivation. This saves an additional call to BLAKE2b and means I can get rid of another class.