src: Use the STREAM construction.
Instead of checking the ciphertext length using associated data in the header, the final chunk is indicated by a different nonce. Thus, the 96-bit counter nonce becomes an 88-bit counter, which is still completely unreachable. This is more flexible and matches what other projects are doing, except incrementing is done in little-endian.

To make the code more readable, I've also removed the zeroing after reading the final plaintext byte. This means any padding is actually plaintext that's already been read into the buffer. Because of how truncation is done, it doesn't matter what the padding bytes are. However, I'll admit this is odd and not ideal.