src: Adjust encrypted header.
This makes the file name header 256 bytes and the spare header 27 bytes. 256 so the padding works with a 255 byte file name. 27 because each timestamp is 8 bytes and only 3 file attributes are worth storing (read-only, hidden, and reparse point). The fourth attribute is whether it's a directory, which is already in use.