EncryptFile.cs: Switch to Covert Encryption's randomised padding scheme.
This implements the padding scheme from the Covert Encryption tool, except a proportion of 10% is used instead of the default 5%. Their tool does a lot of things right, so hopefully they won't mind that I've borrowed some ideas.

I agree with the author(s) that this is better than PADME padding. The previous solution leaked small message lengths, and it seems wrong to have deterministic padding when trying to make files look random. The main issue with this type of approach is if an attacker can request that the same file is encrypted many times.